perf(desktop): split slow tests and frontend chunks
Parallelize chord dictionary coverage and group packaged frontend chunks.
Preserve all 708 tests while removing the oversized bundle warning.

Closes #321
